requirements.txt
Django>=4.0.5
requests>=2.28.0
psycopg2-binary>=2.9.3
django-bulk-update-or-create>=0.3.0